+++ src/playground/configs/templates/PlaygroundPage.vue
@@ -3,6 +3,7 @@
 		<v-row>
 			<ClientTable :settings="tableSettings" />
 			<ServerTable :settings="tableSettingsServer" />
+			<!-- <GroupByClientTable :settings="tableSettings" /> -->
 		</v-row>
 	</v-container>
 </template>
@@ -10,11 +11,12 @@
 
 <script setup>
 import { ref } from 'vue';
+// import GroupByClientTable from './GroupByClientTable.vue';
 import ClientTable from './ClientTable.vue';
 import ServerTable from './ServerTable.vue';
 import tableDefaults from './tableDefaults';
 
-const tableSettings = ref({ ...tableDefaults });
+const tableSettings = ref(Object.assign({}, tableDefaults));
 const tableSettingsServer = ref({ ...tableDefaults, ...{ server: true } });
 </script>