+++ src/plugin/composables/styles.ts
@@ -1,7 +1,6 @@
 import { useConvertToUnit } from './helpers';
 import { useGetLevelColors } from './levelColors';
 import {
-	UseCellStyles,
 	UseHeaderCellStyles,
 	UseTableStyles,
 	UseTFootCellStyles,
@@ -11,13 +10,13 @@ import {
 // -------------------------------------------------- VDrilldownTable //
 export const useTableStyles: UseTableStyles = (options) => {
 	const { colors, level, theme } = options;
-	let baseColors: { border?: string; } = {};
+	let baseColors: { bottomBorder?: string; } = {};
 
 	if (typeof colors === 'object' && colors !== null) {
 		baseColors = useGetLevelColors({
 			colors,
 			level,
-			prop: 'default',
+			prop: 'table',
 			themeColors: theme,
 		});
 	}
@@ -26,8 +25,8 @@ export const useTableStyles: UseTableStyles = (options) => {
 		borderBottom: 'none',
 	};
 
-	if (baseColors.border) {
-		styles.borderBottom = `1px solid ${baseColors.border}`;
+	if (baseColors.bottomBorder) {
+		styles.borderBottom = `1px solid ${baseColors.bottomBorder}`;
 	}
 
 	return styles;
@@ -63,33 +62,8 @@ export const useHeaderCellStyles: UseHeaderCellStyles = (options) => {
 		themeColors: theme,
 	});
 
-	styles.backgroundColor = headerColors.bg;
-	styles.color = headerColors.text;
-
-	return styles as CSSProperties;
-};
-
-
-// -------------------------------------------------- Cell Styles //
-// ! Not used currently //
-export const useCellStyles: UseCellStyles = (options) => {
-	const { colors, elm, level, theme } = options;
-
-	if (colors === false || colors === null) {
-		return {};
-	}
-
-	const baseColors = useGetLevelColors({
-		colors,
-		level,
-		prop: elm,
-		themeColors: theme,
-	});
-
-	const styles = {
-		backgroundColor: baseColors.bg,
-		color: baseColors.text,
-	};
+	styles.backgroundColor = headerColors.background;
+	styles.color = headerColors.color;
 
 	return styles as CSSProperties;
 };
@@ -111,8 +85,8 @@ export const useTFootCellStyles: UseTFootCellStyles = (options) => {
 	});
 
 	const styles = {
-		backgroundColor: baseColors.bg,
-		color: baseColors.text,
+		backgroundColor: baseColors.background,
+		color: baseColors.color,
 	};
 
 	return styles as CSSProperties;